tests() split out tests for vec2.fromScalar fromAngleDegrees fromAngleRadians
