+++ src/core/math/vec2/fromScalar.test.js
@@ -1,0 +1,14 @@
+const test = require('ava')
+const { fromScalar, toString } = require('./index')
+
+const { compareVectors } = require('../../../../test/helpers/index')
+
+test('vec2: fromScalar() should return a new vec2 with correct values', (t) => {
+  const obs1 = fromScalar(0)
+  t.true(compareVectors(obs1, [0, 0]))
+
+  const obs2 = fromScalar(-5)
+  t.true(compareVectors(obs2, [-5, -5]))
+
+  const str = toString(obs2)
+})
